+++ frontend/webpack.config.js
@@ -7,6 +7,8 @@ const TerserPlugin = require('terser-webpack-plugin');
 const OptimizeCSSAssetsPlugin = require("optimize-css-assets-webpack-plugin");
 const VueLoaderPlugin = require('vue-loader/lib/plugin');
 const LicenseWebpackPlugin = require('license-webpack-plugin').LicenseWebpackPlugin;
+const GoogleFontsPlugin = require("@beyonk/google-fonts-webpack-plugin");
+const CreateFileWebpack = require('create-file-webpack');
 
 module.exports = (env, argv) => {
     const devMode = argv.mode !== 'production';
@@ -88,6 +90,23 @@ module.exports = (env, argv) => {
                 filename: devMode ? '[name].[hash].css' : '[name].[chunkhash].css',
             }),
             new VueLoaderPlugin(),
+            new GoogleFontsPlugin({
+                fonts: [
+                    { family: "Source Sans Pro", variants: ["300", "400", "700", "400italic"] }, // cosmo, lumen
+                    { family: "Roboto", variants: ["300", "400", "500", "700"] }, // cyborg, materia, sandstone
+                    { family: "Lato", variants: ["300", "400", "700", "400italic"] }, // darkly, flatly, superhero
+                    { family: "News Cycle", variants: ["400", "700"] }, // journal
+                    { family: "Nunito Sans", variants: ["400", "600"] }, // lux
+                    { family: "Montserrat" }, // minty
+                    { // simplex, spacelab, yeti
+                        family: "Open Sans",
+                        variants: ["300", "400", "700", "300italic", "400italic", "700italic"]
+                    },
+                    //{ family: "Neucha|Cabin Sketch" }, // sketchy
+                    { family: "Source Sans Pro" }, // solar
+                    { family: "Ubuntu", variants: ["400", "700"] }, // united
+                ]
+            }),
         ],
         optimization: {
             runtimeChunk: 'single',
@@ -107,8 +126,11 @@ module.exports = (env, argv) => {
         },
     };
     if (! devMode) {
-        config.plugins.push(new LicenseWebpackPlugin({
-            outputFilename: '[name].[chunkhash].licenses.txt'
+        config.plugins.push(new LicenseWebpackPlugin());
+        config.plugins.push(new CreateFileWebpack({
+            path: '../web/dist',
+            fileName: 'fonts.license.txt',
+            content: 'https://fonts.google.com/attribution'
         }));
     }
     return config;